% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{G1,numeric,character-method}
\alias{G1,numeric,character-method}
\title{one (partly) documented method}
\usage{
\S4method{G1}{numeric,character}(a1, a2)
}
\arguments{
\item{a1}{object of class:\code{numeric}, no manual documentation}

\item{a2}{object of class:\code{character}, no manual documentation}
}
\description{
one (partly) documented method
}
